+++ Sources/Sentry/SentryReplayApi.m
@@ -3,8 +3,9 @@
 #if SENTRY_TARGET_REPLAY_SUPPORTED
 
 #    import "SentryHub+Private.h"
+#    import "SentryOptions+Private.h"
 #    import "SentrySDK+Private.h"
-#    import "SentrySessionReplayIntegration.h"
+#    import "SentrySessionReplayIntegration+Private.h"
 #    import "SentrySwift.h"
 #    import <UIKit/UIKit.h>
 
@@ -22,16 +23,44 @@
 
 - (void)pause
 {
-    SentrySessionReplayIntegration *replayIntegration =
-        [SentrySDK.currentHub getInstalledIntegration:SentrySessionReplayIntegration.class];
+    SentrySessionReplayIntegration *replayIntegration
+        = (SentrySessionReplayIntegration *)[SentrySDK.currentHub
+            getInstalledIntegration:SentrySessionReplayIntegration.class];
     [replayIntegration pause];
 }
 
 - (void)resume
 {
-    SentrySessionReplayIntegration *replayIntegration =
-        [SentrySDK.currentHub getInstalledIntegration:SentrySessionReplayIntegration.class];
+    SentrySessionReplayIntegration *replayIntegration
+        = (SentrySessionReplayIntegration *)[SentrySDK.currentHub
+            getInstalledIntegration:SentrySessionReplayIntegration.class];
     [replayIntegration resume];
+}
+
+- (void)start
+{
+    SentrySessionReplayIntegration *replayIntegration
+        = (SentrySessionReplayIntegration *)[SentrySDK.currentHub
+            getInstalledIntegration:SentrySessionReplayIntegration.class];
+
+    if (replayIntegration == nil) {
+        SentryOptions *currentOptions = SentrySDK.currentHub.client.options;
+        replayIntegration =
+            [[SentrySessionReplayIntegration alloc] initForManualUse:currentOptions];
+
+        [SentrySDK.currentHub addInstalledIntegration:replayIntegration
+                                                 name:NSStringFromClass(SentrySessionReplay.class)];
+    }
+
+    [replayIntegration start];
+}
+
+- (void)stop
+{
+    SentrySessionReplayIntegration *replayIntegration
+        = (SentrySessionReplayIntegration *)[SentrySDK.currentHub
+            getInstalledIntegration:SentrySessionReplayIntegration.class];
+    [replayIntegration stop];
 }
 
 @end